pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "metricstore"
version = "0.1.0"
description = "A lightweight, tool-agnostic metrics catalog with REST API and MCP server"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg",
    "alembic",
    "pydantic>=2.0",
    "pydantic-settings",
    "ruamel.yaml",
    "fastapi-mcp>=0.2",
    "uvicorn[standard]",
    "python-multipart",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
    "aiosqlite",
]

[tool.hatch.build.targets.wheel]
packages = ["src/metricstore"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
